@@ -3,26 +3,28 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [project]
-name = "opencode-debugger"
+name = "python-debugger-mcp"
 version = "0.1.0"
-description = "HTTP relay server enabling AI agents to debug Python applications via debugpy/DAP"
+description = "MCP server for interactive Python debugging with AI agents via debugpy/DAP"
 readme = "README.md"
 license = "MIT"
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 authors = [
-    { name = "OpenCode Team" }
+    { name = "Amir Wilf" }
 ]
-keywords = ["debugger", "debugging", "ai", "dap", "debugpy", "python"]
+keywords = ["mcp", "debugger", "debugging", "ai", "dap", "debugpy", "python", "model-context-protocol"]
 classifiers = [
     "Development Status :: 3 - Alpha",
     "Intended Audience :: Developers",
     "License :: OSI Approved :: MIT License",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
     "Topic :: Software Development :: Debuggers",
+    "Topic :: Software Development :: Testing",
+    "Framework :: FastAPI",
 ]
 
 dependencies = [
@@ -108,12 +113,12 @@ exclude_lines = [
     "if TYPE_CHECKING:",
     "if __name__ == .__main__.:",
 ]
-fail_under = 90
+fail_under = 80
 show_missing = true
 
 # Ruff configuration
 [tool.ruff]
-target-version = "py39"
+target-version = "py310"
 line-length = 100
 src = ["src", "tests"]
 
