@@ -28,24 +28,24 @@ Suggested breakpoints:
 
 from dataclasses import dataclass
 from datetime import datetime
-from typing import Optional
 
 
 @dataclass
 class Order:
     """Represents a customer order."""
+
     id: int
     customer: str
     items: list[str]
     quantities: list[int]
     prices: list[float]
     discount: float = 0.0
-    
+
     @property
     def subtotal(self) -> float:
         """Calculate order subtotal before discount."""
         return sum(q * p for q, p in zip(self.quantities, self.prices))
-    
+
     @property
     def total(self) -> float:
         """Calculate order total after discount."""
@@ -55,7 +55,7 @@ class Order:
 def calculate_statistics(orders: list[Order]) -> dict:
     """
     Calculate statistics for a list of orders.
-    
+
     Try setting a breakpoint here (line 58) to inspect the orders list,
     then step through to watch the statistics being computed.
     """
@@ -66,28 +66,28 @@ def calculate_statistics(orders: list[Order]) -> dict:
         "largest_order": 0.0,
         "items_sold": 0,
     }
-    
+
     # Breakpoint here to watch stats update each iteration
     for order in orders:
         stats["count"] += 1
         stats["total_revenue"] += order.total
         stats["items_sold"] += sum(order.quantities)
-        
+
         if order.total > stats["largest_order"]:
             stats["largest_order"] = order.total
             largest_customer = order.customer  # Try evaluating this!
-    
+
     # Breakpoint here to see final computed stats
     if stats["count"] > 0:
         stats["average_order"] = stats["total_revenue"] / stats["count"]
-    
+
     return stats
 
 
 def process_order(order: Order) -> dict:
     """
     Process a single order and return a receipt.
-    
+
     Step into this function to see order processing details.
     Try evaluating expressions like:
         - order.subtotal
@@ -103,55 +103,57 @@ def process_order(order: Order) -> dict:
         "discount": 0.0,
         "total": 0.0,
     }
-    
+
     # Process each item - good place for conditional breakpoint
     for i, (item, qty, price) in enumerate(zip(order.items, order.quantities, order.prices)):
         line_total = qty * price
-        receipt["items"].append({
-            "name": item,
-            "quantity": qty,
-            "unit_price": price,
-            "line_total": line_total,
-        })
+        receipt["items"].append(
+            {
+                "name": item,
+                "quantity": qty,
+                "unit_price": price,
+                "line_total": line_total,
+            }
+        )
         receipt["subtotal"] += line_total
-        
+
         # Debug tip: Set conditional breakpoint here with "line_total > 50"
         print(f"  {qty}x {item} @ ${price:.2f} = ${line_total:.2f}")
-    
+
     receipt["discount"] = receipt["subtotal"] * order.discount
     receipt["total"] = receipt["subtotal"] - receipt["discount"]
-    
+
     return receipt
 
 
 def find_vip_customers(orders: list[Order], threshold: float = 100.0) -> list[str]:
     """
     Find customers whose orders exceed the VIP threshold.
-    
+
     Good for practicing conditional breakpoints:
         Set breakpoint on line with condition "order.total > threshold"
     """
     vip_customers = []
-    
+
     for order in orders:
         # Conditional breakpoint: order.total > threshold
         if order.total > threshold:
             if order.customer not in vip_customers:
                 vip_customers.append(order.customer)
                 print(f"  VIP: {order.customer} (${order.total:.2f})")
-    
+
     return vip_customers
 
 
 def demonstrate_exception():
     """
     Demonstrates debugging an exception.
-    
+
     Enable "stop on exception" to catch the ZeroDivisionError.
     """
     numbers = [10, 20, 0, 30, 40]
     results = []
-    
+
     for n in numbers:
         try:
             result = 100 / n  # Will raise ZeroDivisionError when n=0
@@ -159,14 +161,14 @@ def demonstrate_exception():
         except ZeroDivisionError as e:
             print(f"  Caught exception: {e}")
             results.append(None)
-    
+
     return results
 
 
 def main():
     """
     Main entry point - demonstrates a complete order processing workflow.
-    
+
     Suggested debugging workflow:
     1. Set breakpoint at line 165 (orders = [...])
     2. Launch and inspect the orders list
@@ -177,7 +179,7 @@ def main():
     print("=" * 60)
     print("OpenCode Debug Demo")
     print("=" * 60)
-    
+
     # Sample orders - inspect this data structure
     orders = [
         Order(
@@ -213,7 +215,7 @@ def main():
             discount=0.15,
         ),
     ]
-    
+
     # Step 1: Calculate statistics
     print("\n1. Calculating order statistics...")
     stats = calculate_statistics(orders)
@@ -222,7 +224,7 @@ def main():
     print(f"   Average order: ${stats['average_order']:.2f}")
     print(f"   Largest order: ${stats['largest_order']:.2f}")
     print(f"   Items sold: {stats['items_sold']}")
-    
+
     # Step 2: Process each order
     print("\n2. Processing orders...")
     receipts = []
@@ -231,20 +233,20 @@ def main():
         receipt = process_order(order)
         receipts.append(receipt)
         print(f"   Subtotal: ${receipt['subtotal']:.2f}")
-        if receipt['discount'] > 0:
+        if receipt["discount"] > 0:
             print(f"   Discount: -${receipt['discount']:.2f}")
         print(f"   Total: ${receipt['total']:.2f}")
-    
+
     # Step 3: Find VIP customers
     print("\n3. Finding VIP customers (orders > $100)...")
     vips = find_vip_customers(orders, threshold=100.0)
     print(f"   VIP customers: {', '.join(vips)}")
-    
+
     # Step 4: Demonstrate exception handling
     print("\n4. Demonstrating exception handling...")
     results = demonstrate_exception()
     print(f"   Results: {results}")
-    
+
     # Summary
     print("\n" + "=" * 60)
     print("Demo complete!")
@@ -252,7 +254,7 @@ def main():
     print(f"Total revenue: ${stats['total_revenue']:.2f}")
     print(f"VIP customers: {len(vips)}")
     print("=" * 60)
-    
+
     return {
         "orders": orders,
         "stats": stats,
